refactor(server): extract session routes
